Data (on pins) is 8bit.

+++ MicroprocessorLib/ReadAbsoluteAddressBank.cpp
@@ -19,8 +19,7 @@ void CReadAbsoluteAddressBank::Init(NotMemoryLock bNotMemoryLock)
 //////////////////////////////////////////////////////////////////////////
 void CReadAbsoluteAddressBank::Execute(CW65C816* pcCPU)
 {
-	//Is this the right Data?
-	pcCPU->GetState()->SetAddressBank((uint8)pcCPU->GetState()->GetData());
+	pcCPU->GetState()->SetAddressBank(pcCPU->GetState()->GetData());
 }
