+++ MicroprocessorLib/ReadDataBank.cpp
@@ -19,8 +19,7 @@ void CReadDataBank::Init(NotMemoryLock bNotMemoryLock)
 //////////////////////////////////////////////////////////////////////////
 void CReadDataBank::Execute(CW65C816* pcCPU)
 {
-	//Is this the right Data?
-    pcCPU->GetState()->SetDataBank((uint8)pcCPU->GetState()->GetData());
+    pcCPU->GetState()->SetDataBank(pcCPU->GetState()->GetData());
 }
